// src/enums.ts
export enum Voice {
  None = 'none',
  Info = 'info',
  Question = 'question',
  Reference = 'reference',
  Example = 'example',
  Story = 'story',
  Lesson = 'lesson',
  Action = 'action',
}

export enum VoiceScope {
  Node = 'node',
  Subtree = 'subtree',
}

export enum PointMovement {
  MoveUp = 'move_up',
  MoveDown = 'move_down',
  IndentLeft = 'indent_left',
  IndentRight = 'indent_right',
}